wip(ia): split team General into Profile/Delete sub-tabs (preserve WIP)
Preserving WIP from cross-session interleaving. Original request:
team /general should follow Members + Policy sub-nav pattern by splitting
the profile form (admin+owner) from the delete flow (owner-only).

Files included:
- /general/page.tsx → server-side redirect to /profile
- /general/layout.tsx → SectionLayout with sub-tabs [プロフィール, 削除]
- /general/profile/page.tsx → name/slug/description form (extracted from
  the original page; admin-or-owner gated)
- /general/delete/page.tsx → owner-only destructive flow with itemized
  warnings (vault / members / audit logs / policies) before the typed-
  confirm AlertDialog
- AdminConsole.json (ja+en) — subTabProfile / subTabDelete keys
- Team.json (ja+en) — section title/desc + delete-warning keys + ownerOnly

Not yet verified (lint/tests/build/pre-pr) due to interrupted session.
Resume with: lint+tests+build → push → PR.
