+++ messages/en/AdminConsole.json
@@ -45,6 +45,8 @@
   "subTabDirectorySync": "Directory Sync",
   "subTabMembersList": "List",
   "subTabMembersTransferOwnership": "Transfer ownership",
+  "subTabProfile": "Profile",
+  "subTabDelete": "Delete",
   "subTabSharing": "Sharing",
   "subTabAccessRestriction": "Access control",
