+++ messages/en/Team.json
@@ -40,6 +40,17 @@
   "deleteTeamTypeHint": "Enter the exact team name \"{name}\".",
   "deleted": "Team deleted",
   "deleteFailed": "Failed to delete",
+  "ownerOnly": "Only the team owner can perform this action.",
+  "generalProfileTitle": "Profile",
+  "generalProfileDesc": "Edit team name, slug, and description.",
+  "generalDeleteTitle": "Delete team",
+  "generalDeleteDesc": "Permanently delete this team and all associated data. This action cannot be undone.",
+  "generalDeleteWarningTitle": "Please review before deleting",
+  "generalDeleteWarningBody": "Deleting \"{name}\" will permanently remove the following data:",
+  "generalDeleteImpactVault": "All passwords and attachments in the team vault",
+  "generalDeleteImpactMembers": "Members, invitations, and role assignments",
+  "generalDeleteImpactAuditLogs": "Audit logs and activity history",
+  "generalDeleteImpactPolicies": "Security policies, webhooks, and encryption keys",
   "invite": "Invite",
   "inviteMember": "Invite Member",
   "inviteEmail": "Email Address",
