+++ messages/ja/Team.json
@@ -40,6 +40,17 @@
   "deleteTeamTypeHint": "チーム名「{name}」と完全一致で入力してください。",
   "deleted": "チームを削除しました",
   "deleteFailed": "削除に失敗しました",
+  "ownerOnly": "この操作はチームのオーナーのみが行えます。",
+  "generalProfileTitle": "プロフィール",
+  "generalProfileDesc": "チーム名・slug・説明を編集します。",
+  "generalDeleteTitle": "チームを削除",
+  "generalDeleteDesc": "チームとすべての関連データを完全に削除します。この操作は取り消せません。",
+  "generalDeleteWarningTitle": "削除前にご確認ください",
+  "generalDeleteWarningBody": "「{name}」を削除すると、以下のデータが復元不能な形で削除されます。",
+  "generalDeleteImpactVault": "チーム保管庫のすべてのパスワードと添付ファイル",
+  "generalDeleteImpactMembers": "メンバー、招待、ロール割り当て",
+  "generalDeleteImpactAuditLogs": "監査ログとアクティビティ履歴",
+  "generalDeleteImpactPolicies": "セキュリティポリシー、Webhook、暗号鍵",
   "invite": "招待",
   "inviteMember": "メンバーを招待",
   "inviteEmail": "メールアドレス",
